Separate out code for testing from production service
