+++ lib/tasks/targets/test_targets.rake
@@ -5,8 +5,8 @@ namespace :targets do
     CSV.open("/tmp/test-targets-report.csv", "w") do |csv|
       csv << ["Group", "School", "Fuel Type", "Enough data", "Recent data", "Success", "Progress", "Error"]
       SchoolGroup.all.each do |school_group|
-        service = Targets::SchoolGroupTargetDataReportService.new(school_group)
-        test_result = service.test_targets
+        service = Targets::SchoolGroupTargetsTestingService.new(school_group)
+        test_result = service.report
         test_result.each do |school, report|
           report.each do |fuel_type, results|
             csv << [
